agentic-sdlc-langgraph/requirements.txt
langgraph
langchain
langchain-openai
streamlit
python-dotenv
lxml
networkx
